backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.12.0
httpx>=0.27.0
python-multipart==0.0.6

# RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
chromadb>=0.4.18
datasets>=2.14.0
sentence-transformers>=2.2.2

